Got the uploading function to work. (#1)
This includes support for authentication via GitHub. Some effort is made to
cache the token for re-use in the same or different R sessions, as well as some
exported functions for token management.

Added a new getter/setter for the REST URL. Updated the REST URL to the new
calcite API, and got a simple example to work.